Added 'speckle' noise
Added 'salt and pepper' noise from the scikit-image library, which is close in meaning to what is called 'specke' noise in the paper.
Still remains to understand whether to run 10 pca on 10 datasets, each with 300 samples of the same number, or to run pca on a big 10*300 samples dataset. This commit does the former, next commit will do the latter (which seems to be the one from the paper).