Implemented the '10*300' dataset version
With the 10 pca on 10 dataset, each with 300 samples of the same number, denoising works perfectly.
With the pca on a single 10*300 dataset with all the numbers, denoising still works fine, but some samples like 2 and 6 get confused as 8. The class 3 example (train from 10*300 samples, with a variable number of features) takes ages to run.